Tests/LibWeb/Text/expected/css/css-image-snapshotted-stylesheet-referrer.txt

Referer is stylesheet URL: true
